Move loading logic into one reusable hook
