+++ packages/changed-elements-react/src/index.ts
@@ -29,11 +29,12 @@ export * from "./contentviews/PropertyComparisonTable.js";
 export type { FilterData, FilterOptions, SavedFiltersManager } from "./SavedFiltersManager.js";
 export { VersionCompareContext, type VersionCompareContextValue } from "./VersionCompareContext.js";
 export * from "./widgets/ChangedElementsWidget.js";
-export * from "./widgets/comparisonJobWidget/versionCompareToasts.js";
 export * from "./widgets/comparisonJobWidget/components/VersionCompareDialogProvider.js";
 export {
-  VersionCompareSelectDialogV2, pollForInProgressJobs,
+  VersionCompareSelectDialogV2,
 } from "./widgets/comparisonJobWidget/components/VersionCompareSelectDialogV2.js";
 export type { JobAndNamedVersions } from "./widgets/comparisonJobWidget/NamedVersions.js";
+export { pollForInProgressJobs } from "./widgets/comparisonJobWidget/useNamedVersionLoader.js";
+export * from "./widgets/comparisonJobWidget/versionCompareToasts.js";
 export { ChangedElementsListComponent } from "./widgets/EnhancedElementsInspector.js";
 export * from "./widgets/VersionCompareSelectWidget.js";
